requirements.txt
fastapi==0.111.0
uvicorn==0.23.2
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
pydantic==2.8.0
requests==2.31.0
python-dotenv==1.0.1
passlib==1.7.4
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
python-multipart==0.0.9